1. The search UI: the gene and phenotype IDs are are now the 2nd and 3rd columns on the search results UI as per @evalkanas feedback (Elise, I will add some question-marks/help for each column soon too, all great points
2. Slack notifications are in place now. All incoming match requests now
will be alerted in #matchmaker_alerts channel. A short description will
be generated for each match by beaker.

3. All incoming match requests will also be logged better in seqr AND
matchbox. The next update will be a seqr page that shows for each
project all the incoming match requests that came in. The backend data
structure for this is now in place.

4. Some other backend updates that should cleanup the backend a little
bit more