Update indeterministic test in test_json_for_saved_variants_with_tags
The original code here does a next(iter(<dictionary>)) which has
an indeterministic order, so this occasionally returns the
variantTag for an AIP variant, which gets transformed with a
different key, causing this test to intermittently fail. For this
reason, we'll hard select the the variant 
VT1726961_2103343353_r0390_100 to look at the keys for.